move zoe swingsetTest helper files up a directory level
Make it accessible to the coming dvp test
